+++ src/Actions/ApiActions.php
@@ -120,17 +120,7 @@ class ApiActions extends DefaultActions
     public function corsOptions($routes = '')
     {
         $origin = $this->requester->getCorsOrigin();
-        if (empty($origin)) {
-            return $this->response();
-        }
-        $this->response = $this->response()
-            ->withHeader('Access-Control-Allow-Origin', $origin)
-            ->withHeader('Access-Control-Allow-Methods', 'GET, POST, OPTIONS')  // PUT, DELETE, PATCH
-            ->withHeader('Access-Control-Allow-Headers', 'X-Requested-With, Content-Type, Accept, Origin, Authorization')
-            ->withHeader('Access-Control-Allow-Credentials', 'true')
-            ->withHeader('Access-Control-Max-Age', '86400')
-            ->withHeader('Vary', 'Origin');
-        return $this->response;
+        return $this->responder->mkCorsOptions($origin);
     }
 
     /**