+++ src/Utilities/ResponseUtil.php
@@ -176,6 +176,26 @@ class ResponseUtil
     }
 
     /**
+     * Create and send CORS options
+     * @param string|false $origin
+     * @return Response
+     */
+    public function mkCorsOptions($origin)
+    {
+        if (empty($origin)) {
+            return $this->response;
+        }
+        $this->response = $this->response
+            ->withHeader('Access-Control-Allow-Origin', $origin)
+            ->withHeader('Access-Control-Allow-Methods', 'GET, POST, OPTIONS')  // PUT, DELETE, PATCH
+            ->withHeader('Access-Control-Allow-Headers', 'X-Requested-With, Content-Type, Accept, Origin, Authorization')
+            ->withHeader('Access-Control-Allow-Credentials', 'true')
+            ->withHeader('Access-Control-Max-Age', '86400')
+            ->withHeader('Vary', 'Origin');
+        return $this->response;
+    }
+
+    /**
      * Create and send the typical OPDS response
      * @return Response
      */