+++ tests/MiddlewareTest.php
@@ -301,6 +301,7 @@ class MiddlewareTest extends PHPUnit\Framework\TestCase
         $request = RequestUtil::getServerRequest('GET', '/admin/');
         $request = $request->withHeader('PHP_AUTH_USER', 'admin')->withHeader('PHP_AUTH_PW', 'admin');
         $middleware = new LoginMiddleware($app->getContainer(), $settings['appname'], []);
+        $middleware->setRequester($request, $settings);
 
         // test protected method using closure bind & call or use reflection
         $isAuthorized = function ($request) {
@@ -322,6 +323,7 @@ class MiddlewareTest extends PHPUnit\Framework\TestCase
         $request = RequestUtil::getServerRequest('GET', '/admin/');
         $request = $request->withHeader('PHP_AUTH_USER', 'admin')->withHeader('PHP_AUTH_PW', 'wrong');
         $middleware = new LoginMiddleware($app->getContainer(), $settings['appname'], []);
+        $middleware->setRequester($request, $settings);
 
         // test protected method using closure bind & call or use reflection
         $isAuthorized = function ($request) {
@@ -343,6 +345,7 @@ class MiddlewareTest extends PHPUnit\Framework\TestCase
         $request = RequestUtil::getServerRequest('GET', '/admin/');
         $request = $request->withHeader('Authorization', 'Basic ' . base64_encode('admin:admin'));
         $middleware = new LoginMiddleware($app->getContainer(), $settings['appname'], []);
+        $middleware->setRequester($request, $settings);
 
         // test protected method using closure bind & call or use reflection
         $isAuthorized = function ($request) {
@@ -364,6 +367,7 @@ class MiddlewareTest extends PHPUnit\Framework\TestCase
         $request = RequestUtil::getServerRequest('GET', '/admin/');
         $request = $request->withHeader('Authorization', 'Basic ' . base64_encode('admin:wrong'));
         $middleware = new LoginMiddleware($app->getContainer(), $settings['appname'], []);
+        $middleware->setRequester($request, $settings);
 
         // test protected method using closure bind & call or use reflection
         $isAuthorized = function ($request) {